requirements.txt
shapeandshare-dicebox-core==3.0.0
pillow==7.2.0
python-mnist==0.7
